kg-mmml/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kg-mmml"
version = "0.1.0"
description = "Integrating Knowledge Graphs with Multi-Modal Machine Learning"
authors = [{name = "Your Name", email = "your.email@university.edu"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "torch>=2.1.0,<3.0.0",
    "transformers>=4.35.0,<5.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "faiss-cpu>=1.7.4",
    "networkx>=3.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--cov=src", "--cov-report=term-missing"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
